Estación-Metereológica/Flask-IoT-Dashboard-Server/requirements.txt
flask
mysql-conector-python
passlib